基本完成查询生成器将生成的代码

+++ src/test/cn/idealismxxm/server/checker/EntityFieldCheckerId.java
@@ -10,12 +10,12 @@ public class EntityFieldCheckerId extends AbstractFieldChecker {
 
     private static final EntityFieldCheckerId CHECKER = new EntityFieldCheckerId();
 
-    private final Class clazz = Integer.class;
+    private final Class clazz = Long.class;
 
     @Override
     public boolean diyCheck(Object value) {
         // 1. 转成真实类型，开始进行 id 字段的定制校验
-        Integer id = (Integer) value;
+        Long id = (Long) value;
 
         // 2. 定制校验
         // TODO your checker code here
